types: move type extension to trpc file
