+++ packages/skill-lesson/trpc/trpc.server.ts
@@ -11,6 +11,26 @@
 import {TRPCError, initTRPC} from '@trpc/server'
 import superjson from 'superjson'
 import {type TrpcContext} from './trpc-context'
+import NextAuth, {DefaultSession} from 'next-auth'
+
+type Resolve<T> = T extends Function ? T : {[K in keyof T]: T[K]}
+
+type User = Resolve<
+  {
+    role?: string
+    purchases?: any[]
+  } & DefaultSession['user']
+>
+
+declare module 'next-auth' {
+  /**
+   * Returned by `useSession`, `getSession` and received as a prop on the
+   * `SessionProvider` React Context
+   */
+  interface Session {
+    user?: User
+  }
+}
 
 const t = initTRPC.context<TrpcContext>().create({
   /**
